+++ htdocs/js/viewsTables.js
@@ -290,8 +290,10 @@ function updateTable(obj, list, perRowCallback, createRowCellsCallback) {
             lastDisc = obj.result.data[i].Disc;
         }
 
-        if (showWorkRow(list) && obj.result.data[0].Work !== undefined &&
-            lastWork !== obj.result.data[i].Work) {
+        if (showWorkRow(list) &&
+            obj.result.data[0].Work !== undefined &&
+            lastWork !== obj.result.data[i].Work)
+        {
             const row = addWorkRow(obj.result.data[i].Work, obj.result.AlbumId, obj.result.Album, colspan);
             if (i + z < tr.length) {
                 replaceTblRow(mode, tr[i + z], row);
@@ -320,27 +322,38 @@ function updateTable(obj, list, perRowCallback, createRowCellsCallback) {
             //default row content
             tableRow(row, obj.result.data[i], list, colspan, smallWidth, actionTd);
         }
-        if (i + z < tr.length) {
-            replaceTblRow(mode, tr[i + z], row);
+        if (features.featPagination === true ||
+            obj.result.offset === 0)
+        {
+            if (i + z < tr.length) {
+                replaceTblRow(mode, tr[i + z], row);
+            }
+            else {
+                tbody.append(row);
+            }
         }
         else {
             tbody.append(row);
         }
     }
-    //remove obsolete lines
-    tr = tbody.querySelectorAll('tr');
-    for (let i = tr.length - 1; i >= obj.result.returnedEntities + z; i --) {
-        tr[i].remove();
+    //remove obsolete rows
+    if (features.featPagination === true ||
+        obj.result.offset === 0)
+    {
+        tr = tbody.querySelectorAll('tr');
+        for (let i = tr.length - 1; i >= obj.result.returnedEntities + z; i --) {
+            tr[i].remove();
+        }
     }
 
     setPagination(obj.result.totalEntities, obj.result.returnedEntities);
-
-    if (obj.result.returnedEntities === 0) {
-        tbody.appendChild(emptyMsgEl(colspan + 1, 'table'));
-    }
     unsetUpdateView(table);
     setScrollViewHeight(table);
-    scrollToPosY(table.parentNode, app.current.scrollPos);
+    if (features.featPagination === true ||
+        obj.result.offset === 0)
+    {
+        scrollToPosY(table.parentNode, app.current.scrollPos);
+    }
 }
 
 /**